refactor: remove unused files and reorganize ImageFetcher
- Delete AlbumImageAdapter, Image, ImageItemInterface, and ImageUrlItem
- Move ImageFetcher to a new directory for better code organization
