--- app/src/main/java/com/example/mydemoapp/adapters/AlbumImageAdapter.java
@@ -1,62 +1,0 @@
-package com.example.mydemoapp.adapters;
-
-import android.content.Context;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import com.bumptech.glide.Glide;
-import com.example.mydemoapp.R;
-import com.example.mydemoapp.models.Image;
-
-import java.util.List;
-
-public class AlbumImageAdapter extends RecyclerView.Adapter<AlbumImageAdapter.ImageViewHolder> {
-
-    private final Context context;
-    private final List<Image> images;
-
-    public AlbumImageAdapter(Context context, List<Image> images) {
-        this.context = context;
-        this.images = images;
-    }
-
-    @NonNull
-    @Override
-    public ImageViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(context).inflate(R.layout.item_image, parent, false);
-        return new ImageViewHolder(view);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull ImageViewHolder holder, int position) {
-        Image image = images.get(position);
-
-        Glide.with(context)
-                .load(image.getPath())
-                .into(holder.imageView);
-    }
-
-    @Override
-    public int getItemCount() {
-        return images.size();
-    }
-
-    public static class ImageViewHolder extends RecyclerView.ViewHolder {
-        ImageView imageView;
-
-        public ImageViewHolder(@NonNull View itemView) {
-            super(itemView);
-            imageView = itemView.findViewById(R.id.image_view);
-        }
-    }
-
-    public void setImages(List<Image> images) {
-        this.images.clear();
-        this.images.addAll(images);
-    }
-}
