--- app/src/main/java/com/example/mydemoapp/models/Image.java
@@ -1,13 +1,0 @@
-package com.example.mydemoapp.models;
-
-public class Image {
-    private String path;
-
-    public Image(String path) {
-        this.path = path;
-    }
-
-    public String getPath() {
-        return path;
-    }
-}
