--- app/src/main/java/com/example/mydemoapp/models/ImageItemInterface.java
@@ -1,7 +1,0 @@
-package com.example.mydemoapp.models;
-
-public interface ImageItemInterface {
-    String getDate();
-    String getImageUrl();
-    int getImageId();
-}
